+++ tests/ka/test_range.cpp
@@ -343,7 +343,7 @@ TEST_P(RangeBoundedRangeLinearizableBeginEnd, Basic) {
   EXPECT_EQ(b, begin(rng));
   EXPECT_EQ(e, end(rng));
   EXPECT_EQ(values.empty(), empty(rng));
-  EXPECT_EQ(values.size(), size(rng));
+  EXPECT_EQ(values.size(), static_cast<std::size_t>(size(rng)));
   for (auto i = 0; i != size(rng); ++i) {
     EXPECT_EQ(values[i], rng[i]);
   }
